navigation and integration changes and ground assessor app bug fix
